requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.3.0
better-profanity==0.7.0
bidict==0.23.1
bitarray==3.3.1
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.2
cattrs==24.1.3
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.2.3
comtypes==1.4.11
contourpy==1.1.1
cryptography==44.0.2
cssselect2==0.7.0
cycler==0.12.1
cytoolz==1.0.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
eventkit==1.0.3
exceptiongroup==1.3.0
executing==2.2.1
Faker==35.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SocketIO==5.5.1
Flask-WTF==1.2.1
flatbuffers==25.2.10
fonttools==4.57.0
fpdf2==2.8.3
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gast==0.4.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gevent==24.2.1
google-auth==2.42.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.16.0
h5py==3.11.0
hexbytes==1.3.0
html5lib==1.1
idna==3.10
imageio==2.35.1
imbalanced-learn==0.12.4
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
inflection==0.5.1
intersect==1.2
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
jh2==5.0.9
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
keras==2.13.1
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.41.1
lxml==5.3.2
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
mysql-connector-python==9.0.0
mysqlclient==2.1.1
nest-asyncio==1.6.0
networkx==3.1
niquests==3.14.1
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
oscrypto==1.3.0
packaging==24.2
pandas==2.0.3
parsimonious==0.10.0
parso==0.8.5
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prompt_toolkit==3.0.52
propcache==0.2.0
protobuf==4.25.8
psutil==7.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2024.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qh3==1.5.2
ratelim==0.1.6
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
simple-websocket==1.1.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.4.0
toolz==1.0.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
urllib3-future==2.12.922
wassima==1.2.2
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.0.6
wrapt==2.0.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.15.2
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
